// fe/src/types/generated/RoleResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PermissionResponse } from "./PermissionResponse";

/**
 * Role response DTO
 */
export type RoleResponse = { 
/**
 * UUID of the role
 */
uuid: string, 
/**
 * Name of the role
 */
name: string, 
/**
 * Description of the role
 */
description: string | null, 
/**
 * Whether this is a system role
 */
is_system: boolean, 
/**
 * Whether this role grants super admin privileges
 */
super_admin: boolean, 
/**
 * Direct permissions for this role
 */
permissions: Array<PermissionResponse>, 
/**
 * When the role was created
 */
created_at: string, 
/**
 * When the role was last updated
 */
updated_at: string, 
/**
 * UUID of the user who created the role
 */
created_by: string, 
/**
 * UUID of the user who last updated the role
 */
updated_by: string | null, 
/**
 * Whether the role is published
 */
published: boolean, 
/**
 * Version number
 */
version: number, };