// fe/src/types/generated/WorkflowSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type WorkflowSummary = { uuid: string, name: string, kind: string, enabled: boolean, schedule_cron: string | null, 
/**
 * Indicates if this workflow has a from.api source type (accepts POST, cron disabled)
 */
has_api_endpoint: boolean, versioning_disabled: boolean, };