// fe/src/types/generated/WorkflowDetail.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type WorkflowDetail = { uuid: string, name: string, description: string | null, kind: string, enabled: boolean, schedule_cron: string | null, config: unknown, versioning_disabled: boolean, };