// fe/src/types/generated/WorkflowRunSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type WorkflowRunSummary = { uuid: string, status: string, queued_at: string | null, started_at: string | null, finished_at: string | null, processed_items: number | null, failed_items: number | null, };